Move wrappers to separate file
